✨ Create a new component with the editing behavior
